Code downloads as UTF8 and wont open correctly
Hi Chris, hope you are well. Had issue downloading this today in that it opens in ISE Editor, but in Notepad appears UTF8 with spaces between each character and does not run in the console. I have just pasted the code back in to generate the change, but there might be a better way to fix the download.

Brian